requirements.txt
# Flask and web server components
Flask
gunicorn
eventlet

# MQTT client
paho-mqtt

# HTTP requests
requests

# Google Gemini AI
google-generativeai

# For video processing (used by main.py indirectly via ffmpeg)
# No package needed here, ffmpeg is installed via apt in the setup script.
